Reorganize to work will with coq_makefile
Now the *.v files are all in the subdirectory Foundations, and "make install"
will install them in the directory user-contrib/Foundations inside the
currently running coqc distribution tree.

Instead of using coq_makefile to make the file "Makefile", we use it to make
the file "Make.makefile".  That way we have a Makefile we can edit manually,
which is convenient, for example, for making a TAGS file.