add loadpath commands for interactive "Foundations" access
Now that we are using fully qualified library names, such as
Foundations.Generalities.uu0, we have to arrange for coqtop to work, especially
when run by Proof General.  One way would be to give it "-R . Foundations" as
command line options.  But Proof General has a strange sort of bug that causes
it to loop forever when starting up coq after you evaluate

     (setq coq-prog-args '("-emacs" "-R" "." "Foo"))

.  The best solution seems to be to add lines

      Add LoadPath ".." as Foundations.

to our *.v source files, which we do in this commit.