comment out the recently added lines  Add LoadPath ".." as Foundations
The reason is that the path is relative, but not relative to the directory
containing the source file, but rather, relative to the current working
directory of the compiler.  This can cause confusion if a sibling to the top
level directory of this repository happens to be called "Foundations", because
coq_makefile generates compilation commands that are run in the top level
directory.