found an even better definition of isdecprop
... namely:

  Definition isdecprop (P:UU) := isaprop P × (P ⨿ ¬P).

Showing that isdecprop(P) is a proposition will be done with funextfun.

We avoid asserting contractibility of P ⨿ ¬P because that can only be proven
with the aid of funextempty.

We avoid providing a replacement proposition for ¬P as part of the data,
because then isdecprop(P) is in a higher universe, and because we can easily
get a new proposition Q logically equivalent to ¬P:

  Definition neg_decprop {P} (i: isdecprop P) := Σ n:¬P, ii2 n = pr2 i.

Proving Q is a proposition will not use funextempty, because it will depend
only on ii2 being an inclusion:

  Lemma isaprop_neg_decprop {P} (i:isdecprop P) : isaprop (neg_decprop i).
    (to be proven)

Proving the logical equivalence will:

  Lemma neg_decprop_negates {P} (i:isdecprop P) : neg_decprop i <-> ¬ P.