[colimits]: add a few lemmas, one failing
The one without a trivial proof is this:

Lemma functorBinarySum_map_eqn {B C:Precategory} (sum : hasBinarySums C)
      (F G F' G' : [B,C]) (p:F→F') (q:G→G') (b:B) :
  binarySumMap_2 (functorBinarySum sum F G) (functorBinarySum sum F' G') p q ◽ b
  =
  binarySumMap_2 (sum (F ◾ b) (G ◾ b)) (sum (F' ◾ b) (G' ◾ b)) (p ◽ b) (q ◽ b).

This failure might be what prevents using this framework with
SubstitutionSystems on the branch "colimits". Since
[functorBinaryProduct_map_eqn] admits a trivial proof, that's an argument for
replacing the proof of functorBinarySum by one that's parallel to the proof of
functorBinaryProduct, rather than deducing it as a corollary.  Maybe then we
could also write [universalObject sum] instead of [rm_opp_ob (universalObject
sum)]