Update coq to latest
We update Coq to the latest 8.5 version after the announcement of
v8.5 RC 1.

This update implements "simple refine", which is the recommended way to
get the new behavior of the "refine" tactic, so we replace "unshelve
refine" by "simple refine" everywhere except in those cases where it
doesn't work.  (It was just a few days ago that we had to replace almost
all of our uses of "refine" by "unshelve refine".)