fix notations
Resolves issue https://github.com/UniMath/UniMath/issues/322 .

We changed the level of Notation "x → y" from 90 to 99 to match the core
of Coq, and we removed the Notation "f ;; g" from the Ktheory package.

I don't understand exactly what provokes this incomptability -- there
are certainly some bugs in the way Coq takes care of such things.  In
particular, we aren't reserving any notations yet -- it's a good
suggestion, especially if notation reservation within a given scope
would work.