change notation for "forall"
We rename our notation for ∀ to Π.

We make a global definition for ∀ that applies just to families of
propositions and yields a proposition (hProp).

We change all existing uses of "forall" and ∀ to Π.  Later on, those
that applied to hProps could be switched back to ∀.

The reason we change all uses of "forall" is that ∀ is normally read as
"for all".

Here is the definition of ∀ from Propositions.v:

    Lemma isaprop_forall_hProp (X:UU) (Y:X->hProp) : isaprop (Π x, Y x).
    Proof. intros. apply impred_isaprop. intro x. apply propproperty. Defined.

    Definition forall_hProp {X:UU} (Y:X->hProp) : hProp := hProppair (Π x, Y x) (isaprop_forall_hProp X Y).

    Notation "∀  x .. y , P" := (forall_hProp (fun x => .. (forall_hProp (fun y => P)) ..))
      (at level 200, x binder, y binder, right associativity) : type_scope.