clean up in preparation for primitive projections
The use of "induction" or "destruct" on an element of type "total2 P"
without explicit names for the new variables will result in invented
names such as t, p, t0, p0, ..., but those will change to pr1, pr2,
... when we switch to primitive projections.  We locate fix all such
usage by renaming the two parts of total2.  We do something similar for
elements of type 'coprod A B'.