move some results in PartC earlier
In the discussion about pull request https://github.com/UniMath/UniMath/pull/257 Vladimir
suggested (see https://github.com/UniMath/UniMath/pull/257#issuecomment-231534974 ) that results in
PartC not depending on `funextempty` be moved to PartB.  I took that as a request to do so in #257,
but that made the pull request much more complicated.  Here I move those results in a single commit,
intended to be part of a single pull request.  I anticipate this will make a future version of #257
simpler.

We fix the comment at the start of PartC.v to inform future editors of the file about the policy.