primitive projections: fix Archimedean.v better
We install a new tactic so you can choose what to do during a proof
based on whether primitive projections are enabled.  This is important,
because with primitive projections enabled, the parameters to "pr1" and
"pr2" are not present, and thus index numbers of subexpressions can
change, as happened in this case.