decidable proofs
When trying to prove a statement that is decidable, it is not necesary
to think about the proof, because the proof of decidability implements
an algorithm for deciding.  All we need to do is to figure out how to
"run" it.  The same applies to proving the negation of such a statement.

Here we offer some tactics that make the notation for doing so easy.

The idea is similar to the "canonical term" notation "ct", and we put it
in the same place in the code.

We illustrate the method by cleaning up 5 proofs about rational numbers
in RealNumbers/Prelim.v, each of which seemed ad hoc.  The new proofs
seem more uniform and involve no thought when constructing them.