remove submodule lablgtk
Resolves #383.

It turns out that there is a homebrew "tap" called "mht208/formal"
that contains a recipe for building "ocamlfind", which is now required
by the makefile that builds "coq".  Here we take advantage of that to
remove the submodule "lablgtk", saving some time.

Note that "lablgtk" is used only if we build coqide.