remove lemma apevalat
In pull request #527 Vladimir introduced the function `eqtohomot` to replace
`apevalat` partially.  In this pull request I eliminate `apevalat` in favor of
`eqhomotop`, generalizing `eqtohomot` from functions to sections to make that
possible.  I also moved the definition of `toforallpaths` to PartA, to point
out that the only difference between it and `eqtohomot` is the number of
implicit arguments.

Probably we should eliminate `toforallpaths` in favor of `eqtohomot` later on.