enforce linear ordering of proof files
We have a linear ordering of the proof files, since the packages are listed in
the file `Makefile` and the proof files within a package are listed in
`.package/files`.  The linear ordering is used for making the `TAGS` file.

It's a good idea for the linear ordering to be compatible with the dependency
ordering, because then the global organization of UniMath is more
understandable, and also a search using tags will progress through the files in
a logical ordering (e.g., a search for a global identifier will find its
definition first).

In this commit we add some lines to `Makefile` to enforce the compatibility.
This depends on the shell being `bash`, but for all practical purposes, `bash`
has replaced more naive Bourne shells, so this dependency is safe.