open branch decidable-propositions
We want to fix "choice" so it works with all decidable propositions.
See the items in Combinatorics/Tests.v marked "fix".  The fix is
to remove the use of funextempty by replacing the old definition of
decidability involving negation by the notion of complementary pairs
and the two propositions (isTrue and isFalse) that can be derived from
them.