binary operation iteration over unordered sequences
Got this (finally):

 Definition iterop_finseq_mon {M:abmonoid} : UnorderedSequence M -> M.
 (* iterate the monoid operation over a finite sequence of elements of M *)

An unordered sequence in M is a function to it from a finite set.  The proof
uses the generalized commutativity theorem (a finite product is independent of
the ordering).  The proof of commutativity uses the generalized associativity
theorem (a product of products is a product).  I almost had this proof more
than a year ago.