check for any unlisted proof files
The makefile now ensures that each *.v file inside a package directory
is listed in the corresponding file UniMath/*/.package/files.  A single
unlisted file is permitted, to give the author time to look at it and
decide what to do about it: CategoryTheory/equivalences_lemmas.v .  The
unlisted file Ktheory/Sets.v was removed.

resolves #723