tactic for proving that types have some hlevel
This tactic is usful for proving goals of the form isofhlevel n X.
It dissects commonly used type formers such as direct product,
sigma, and pi, and then tries to analyze their components.

It reduces the burden on the programmer to remember the names of
such tactics, and to figure out which ones should be applied, in which
branches, how many times, and in what order.

There is plenty of room for improvement, but this is a solid first step.