stop using Coq's initial prelude
This resolves issue #827 partially.  We still load Init.Logic, because
that extends the capability of the tactic "rewrite" in useful ways that
we use, but it introduces terms that use Prop, which we don't want.  (I
spent some time replacing failing uses of "rewrite" by other things, but
tired of that.)

We also still load Init.Notations, but that is innocent, because we want
to use the standard Coq notations reserved there.

I'm currently puzzled about how to get parsing of numerals, such as "4",
to work, so as a stop-gap measure, I've defined notations for the
numerals less than 25.

I've also renamed the tactic "unimath_easy" to "easy", since Coq's
"easy" is no longer present.