fix init
We address issue #827.

This commit adds a file Coq/Init/Datatypes.v to UniMath in order to make
the natural number syntax plugin work once again -- it hard codes access
to the fully qualified identifiers Coq.Init.Datatypes.O and
Coq.Init.Datatypes.S, and can't work without those.

But I see a problem: after "make install" our copy of Coq.Init.Datatypes
will be preferred to Coq's copy, whenever using the same Coq
installation:

... if I use "-noinit":

    Datatypes.vo has been found in
    [ /Users/dan/src/ProofChecking/UniMath-master/sub/coq/user-contrib/Coq/Init ;
      /Users/dan/src/ProofChecking/UniMath-master/sub/coq/theories/Init ];
    loading /Users/dan/src/ProofChecking/UniMath-master/sub/coq/user-contrib/Coq/Init/Datatypes.vo
    [ambiguous-file-name,filesystem]</warning>
    <infomsg>[Loading ML file ltac_plugin.cmxs ...
    [Loading ML file nat_syntax_plugin.cmxs ...
    </infomsg>

It loads our version and it works.

... if I don't use "-noinit"

    Welcome to Coq 8.7.1 (December 2017)
    Error during initialization:
    Error:
    Compiled library Coq.Init.Prelude (in file /Users/dan/src/ProofChecking/UniMath-master/sub/coq/theories/Init/Prelude.vo) makes inconsistent assumptions over library Coq.Init.Datatypes

So installation will spoil the use of that Coq for anything but UniMath.

Alternatively, if we copy the file nat_syntax.ml and adapt it, we can
fix that problem, but then installation of UniMath on a system that
already has Coq will also require ocaml to be there, as well as a step
where nat_syntax.ml is compiled and the compiled file is placed
somewhere findable (?).

I think these limitations are inherent in the current situation.