fix the Makefile
After updating the makefile so it would no longer include the makefile built by
coq_makefile, it would no longer correctly determine which style for dependency files is in use.
We arrange for the *.d files to be made, because we need to read them to enforce the prescribed ordering.
Up to coq version 8.7, each *.v file had a corresponding *.v.d file.
After that, there is just one *.d file, it's name is .coqdeps.d, and it sits in this top-level directory.
So we have to distinguish the versions somehow; here we do that by grepping the makefile built by
coq_makefile for VDFILE.