remove redundant properties from definition of direct sum
We replace the access functions for the redundant properties with
proofs of them.

As a side effect of adding about 100 lines of proof, we are able to
remove 630 lines of redundant proofs devoted to proving the redundant
properties.

The old definition:

     Definition isBinDirectSum (a b co : A) (i1 : a --> co) (i2 : b --> co)
		(p1 : co --> a) (p2 : co --> b) : UU :=
       (isBinCoproduct A a b co i1 i2)
	 × (isBinProduct A a b co p1 p2)
	 × (i1 · p1 = identity a) × (i2 · p2 = identity b)
	 × (i1 · p2 = (to_unel a b)) × (i2 · p1 = (to_unel b a))
	 × ((to_binop co co) (p1 · i1) (p2 · i2) = identity co).

The new definition (with some new notations) :

     Definition isBinDirectSum (a b co : A) (i1 : a --> co) (i2 : b --> co)
		(p1 : co --> a) (p2 : co --> b) : hProp :=
       i1 · p1 = 1  ∧  i2 · p2 = 1  ∧
       i1 · p2 = 0  ∧  i2 · p1 = 0  ∧
       p1 · i1 + p2 · i2 = 1.

To make the parsing of the new expression work, we take the opportunity
to make the operator f·g bind as tightly as the other multiplication
operations -- that's the single change to Foundations.

On branch fix-binary-direct-sums
Changes to be committed:
	modified:   UniMath/CONTENTS.md
	modified:   UniMath/CategoryTheory/AbelianToAdditive.v
	modified:   UniMath/CategoryTheory/AdditiveFunctors.v
	modified:   UniMath/CategoryTheory/PrecategoriesWithAbgrops.v
	modified:   UniMath/CategoryTheory/categories/abgrs.v
	modified:   UniMath/CategoryTheory/limits/BinDirectSums.v
	modified:   UniMath/Foundations/Init.v
	modified:   UniMath/HomologicalAlgebra/Complexes.v
	modified:   UniMath/SubstitutionSystems/SubstitutionSystems.v