exact categories
Here we fix up our use of algebraic notations so there is less
ambiguity (having two notations for the same thing) and so one can avoid
certain notations by not loading the module containing them.

The goal was to make display and input of terms convenient and concise,
but it wasn't quite achieved.  There are some comments explaining the
problems.