delete some files that use inductive constructions; move a file
Overview of deleted files
=========================

Construction of the product of a family of sets with a binary operation (magmas).

The zero monoid is a zero object.  Construction of monoids by generators and
relations.  Construction of free monoids.  Construction of a product of
monoids.  Uses an inductive construction.

The zero abelian monoid is a zero object.  Construction of abelian monoids by
generators and relations.  Construction of free abelian monoids.  Construction
of a product of abelian monoids.  Forthcoming: the sum of a finite family of
elements.  Uses an inductive construction.

The zero group is a zero object.  Construction of groups by generators and
relations.  Construction of free groups.  Construction of a product of groups.
Uses an inductive construction.

The zero abelian group is a zero object.  Construction of abelian groups by
generators and relations.  Construction of free abelian groups.  Construction
of products and sums of abelian groups.  The category of abelian groups, with
abstract products and sums.  Uses an inductive construction.