Add braided monoidal precategories
Defines braidings and the braided monoidal structure on top of them.
They are required to satisfy the hexagon equations. Accessors and
a coercion from braided monoidal to monoidal are provided.

In #722, braided monoidal categories were discussed, but the pull
request was too big to be merged. In #1009, monoidal categories were
added, but braided monoidal categories were not part of the pull
request.

Note that this adds only braided monoidal PREcategories, even when
the name of the file tries to match MonoidalCategories.v, which only
provides precategories.

Also update .package/files.