Monoidal categories as bicategories
Every monoidal category gives rise to a bicategory where the type of
0-cells is the unit type.  We check that all the equations on the
definition of a prebicategory are satisfied by the monoidal precategory.

The only difficult one is the pentagon equation, that reads backwards
on one of the two definitions.

It should be easy to extend this to monoidal categories and
bicategories.  The hope would be that we can use some of the general results
from bicategories in monoidal categories without having to reprove
them again.