WIP towards an isomorphism of categories in ActionBasedStrengthOnHomsInBicat.v
Two lemmas currently admitted - help needed to assess if there is any chance to prove those lemmas. They ask for proving equality of morphisms in a given category, involving Univalence.double_transport.

a number of preparations for that development in the other files:
laws for ab strengths and signatures are props
not only the precategories but also the categories of signatures and of ab strong functors are defined