tinkering to solve a problem - no results yet
BicatOfCatsWithoutUnivalence.v: the very end shows efforts to get univalence of bith involved categories, the second one presents major problems with non-terminating type-checking. The notion of action seems to depend on assumed univalence of the base category.

For test purposes, I made transparent two lemmas on homsets in functor categories: in the elemenary and the bicategorical setting.