identifies developments that can be done with the bicategory of categories (CAT)
This means, without taking univalent categories (UCAT) but just precategories with homsets.

The new files are simple adaptations of the respective files without "WithoutUnivalence" in their name.

Biadjunction.v now does not refer to UCAT but to CAT, this needed more material in BicatOfCatsWithoutUnivalence.v to be taken from the development for UCAT.

Final.v and Initial.v did not even depend on UCAT (a redundant import)
Strictness.v only depends on UCAT for an example (better, a counter-example)
Monads.v: disp_2cells_isaprop_monad does not need the hypothesis is_univalent_2 C
MonadKtripleBiequiv.v: adapted to that change in Monads.v