identifies inserter category as abstract construction behind final section of CPP'22 paper
The basic situation is now developed in CommaCategories.v, with the equation being a special case of the one for comma categories - which is turned around w.r.t. to the previous definition. Therefore, in the application, some equations were turned around in the statements, but mostly, there are now more uses of pathsinv0.