Remove hset_fun_space.
hset_fun_space is a duplicate of funset, defined in MoreFoundations/PartA.v

This also changes the type of morphisms in HSET. Before this commit
the morphisms were defined to be hSets, but there is no reason to bundle the
isaset-property with each homset.

This commit defines the homsets in HSET to be the ordinary function type.