fix issue 1577
resolves #1577

We add the build target "html" when building "all" so the *.glob files will be built.
Without that, the install step will fail.
See the issue at https://github.com/UniMath/UniMath/issues/1577.
It's a little bit sad that we have to make the html files always, but the dependencies in Coq's makefile, build/CoqMakefile.make, are not right:
the install target doesn't depend on the glob files, even though it needs them.