Remove notation ';;' and scope [cat_deprecated] from Categories.
The notation ';;' for composition of morphisms is rarely used, and some of the
few places where it was used was in DispCats where ';;' is typically used for
composition of _displayed_ morphisms_. To avoid confusion this notation is
removed.

The scope [cat_deprecated] is empty and unused following this change and
removed.