CI: Fix a source of some cache misses.
In rare circumstances the build would miss the cache. For example, this would
happen when a job in a pull request compiled using coq-dev had failed on a
previous run, but coq-latest succeeded. Then the current job would restore the
coq-latest cache. This cache is completely useless when compiling with
coq-dev. Instead it should restore any available coq-dev cache from the
master-branch.

This change makes the restore function only match caches with the same
coq-version.