CI: Remove superfluous cache files.
Currently each cache run will create a new copy of the cache files. When a PR is
being worked on with many changes happening this can quickly use up the quota of
cache space available.

With this change old cache files will be removed before the new ones are saved,
essentially making each run update the cache instead of storing an updated copy.