Add a definition for algebras of a clone and prove some results about algebras (#1681)
I believe the equivalence between algebras for the free monoid clone and
monoids was my hardest proof until now. I feel like the result is
okay-ish, but I had to do a lot of manual rewriting.
* A couple of days ago, I discovered `rewrite (idpath _ : term_a =
term_b)` which I consider to be better than a lot of unfolding and
refolding, but I am not sure whether this is considered good or bad
practice.
* The names in this package have become quite long, so I had to insert
some (relatively arbitrary) line breaks. I am not sure whether there are
any style guides about how to format long lines (except the requirement
that lines be 100 characters or less).