Formalized notion of model of enriched effect calculus (#1695)
I used the formalization on enriched categories to formalize Definition
6.5, Proposition 6.9, and Example 6.10 from
http://www.itu.dk/people/mogel/papers/eec.pdf.

Note: I created a new folder for this, because I think that in the short
run, there will be more stuff added to that folder. In addition, I am
hesitant of putting it in CategoryTheory, because I would like to be
able to import Bicategories for these files (although that currently is
not done, yet). As such, the two options were to either add it to
Bicategories or to create a new folder for it. I found the second to be
more logical.