Refactor sub bicategories (#1708)
In this PR:
- The notion of subbicategory (`subbicat`) is made more general
- I added two examples displayed bicategories over the bicategory of
categories (not necessarily univalent ones). Both of these are about
terminal objects: one is for chosen terminal objects, the other is for
existing terminal objects.

The main reason to look at these bicategories is to study how it
interacts with the Rezk completion. More specifically, what can we say
about the Rezk completion of a category with a terminal object? Does it
matter whether the terminal object is chosen or just exists?