The Grothendieck construction as a biequivalence (#1735)
This PR contains a proof of the Grothendieck construction as a
biequivalence of bicategories (see
https://ncatlab.org/nlab/show/Grothendieck+construction#statement_of_the_equivalence).

There currently is 1 hole in the proof. I make use of univalent
bicategories, and it still needs to be proven that the bicategory of
fibrations over a fixed category `C` is univalent. This is for a later
PR.