[WIP] Double categories via 2-sided displayed categories (#1720)
This PR contains an alternative definition of double categories,
formulated with 2-sided displayed categories. In addition, it contains a
function that maps this definition of double categories to the usual one
(as formalized by Nima some time ago). It is worthwhile to note that on
the data, it is the identity, but that only the laws are formulated a
bit differently.