Bicategory of univalent double categories (#1751)
This PR is currently work in progress. The content is going to be:
- The bicategory of univalent double categories (defined via a displayed
bicategory)
- A proof that this bicategory is univalent
- Refactor the notion of double category (in the directory on 2-sided
displayed categories) accordingly, so that it is the same to the
obtained notion of double category via the construction of this
displayed bicategory.

This required a refactoring in the directory on 2-sided displayed
categories to make certain proofs more efficient (regarding memory) and
simpler.