Continuous DCPO structure (#1762)
In Tom de Jong's thesis (https://tdejong.com/writings/phd-thesis.pdf),
several variations of continuity for DCPOs is given (Chapter 4), and the
difference between them is whether they are structure or property. In
this PR, I show that a continuity structure for a DCPO can be chosen if
one knows it exists (without assuming any choice axiom). The main idea
behind this proof is that there is a canonical choice, and this gives us
a choice.

For algebraic DCPOs, a similar proof is given.