Orthogonal factorization systems in bicategories (#1794)
This PR cleans the file on eso 1-cells in bicategories. It does so by:
- defining the notion of orthogonality for 1-cells of bicategory (this
means that certain lifting properties are satisfied)
- refactor the eso file by basing it on orthogonality
- defining the notion of orthogonal factorization systems in
bicategories
- constructing the (eso, ff) factorization system for 1-types and for
univalent categories

I am planning to also construct an orthogonal factorization system for
enriched categories (using eso and fully faithful functors), but that's
for a separate PR. From this factorization system, we directly get that
functors that are essentially surjective and fully faithful, also are
adjoint equivalences.