Universal property of the Rezk completion for enriched categories (#1797)
This PR builds forth on https://github.com/UniMath/UniMath/pull/1796.

It contains a proof of the universal property of the Rezk completion of
enriched categories. The main idea behind the proof is the same as for
ordinary categories. The only difference is that the enrichments must be
constructed as well, and those can be dealt with using the same methods.