Refactoring enriched categories, work on Kleisli categories (#1798)
This PR builds forth upon https://github.com/UniMath/UniMath/pull/1797.

This PR is a bit big. It contains three refactorings:
- To prove that a transformation is enriched,
`nat_trans_enrichment_via_comp` is chosen, because this simplifies the
proofs.
- The composition operator in the Kleisli category and Eilenberg-Moore
category is slightly changed. Now `precomp_arr` and `postcomp_arr` are
used systematically, because this simplifies the proofs. This can
especially be noted in the file on Kleisli categories. Material is added
about the universal property of the non-univalent Kleisli category, but
the file became significantly smaller.
- I finally proved that an enriched transformation is invertible if and
only if its underlying transformation is invertible. Before, I didn't
use `nat_trans_enrichment_via_comp` and that made the proof harder. Now
the proof is rather simple.

I also added the functors and transformations for the universal property
of the non-univalent Kleisli category.