Lambda calculus and Scott's Representation Theorem (#1786)
This PR formalizes a part of Hyland's 2013 paper "Classical lambda
calculus in modern dress". It adds lambda theories, presheaves for
algebraic theories, their categories (including univalence and limits),
some tools in the library (mainly for defining the categories, working
with them, and proving that they have limits), and ultimately shows
Scott's Representation Theorem (and so it contains all results and
definitions needed by this theorem).
The changes in the `AlgebraicTheories` package can be categorized as
follows:
* Basic definitions: `AlgebraicTheor(ies2?|yMorphisms2?|yCategory)`
* The same for `Algebra`, `Presheaf` en `Lambda`.
* Some changes and updates to `Examples`, which are examples and
standard constructions for these definitions.
* Some miscellaneous files: `FiniteSetSkeleton`, `LambdaAlgebraMonoid`,
`LambdaCalculus`, `Tuples`.
* `RepresentationTheorem`. Here (almost) everything comes together to
prove the representation theorem.