More ondouble categories (#1807)
I added some basic infrastructure and examples of strict double
categories. It is good to note that these examples arise from strict
categories, because that's just how strict double categories are.

I also added a symmetric univalence condition for pseudo double
categories, and an example of such a double category.