Refactor univalent double categories (#1810)
This is another refactoring on double categories. It separates a notion
of double category that is close to the definition based on 2-sided
displayed categories (so that that material can be reused), but without
a univalence condition. As such, several proofs/constructions for pseudo
double categories become reusable for both the univalent and the strict
case (horizontal bicategory, vertical 2-category, some laws). The
disadvantage is that at some points, one needs to tell more to the type
checker of Coq, which is why `(C := C)` got added at several lines.