Refactoring 2-sided displayed categories and duality involutions (#1813)
This PR contains some refactorings:
1. The notion of 2-sided displayed category only depends on the data of
the categories over which it lives. Because of this change, at several
points I needed to tell the relevant 2-sided displayed category
2. I created separate files for properties of pseudofunctors, and I
looked at some properties of local equivalences. I also renamed the
notion of essential surjectivity for pseudofunctors to prevent a name
clash.
3. The notion of duality involution got refactored, and some additional
properties got proven. I added documentation, and I simplified one of
the coherences, so that it becomes simpler to use/prove.

The first refactoring is a preparation for a formalization of double
bicategories a la Verity. The purpose of the third refactoring is to
look at the interaction of duality involutions and Eilenberg-Moore
objects, which generalizes the fact that the Eilenberg-Moore category
for a comonad can be constructed from the Eilenberg-Moore category for a
monad.