The biequivalence by Clairambault and Dybjer (#1857)
This PR contains one of the biequivalences by Clairambault and Dybjer.
More specifically, I construct a biequivalence between the bicategory of
univalent categories with finite limits and univalent comprehension
categories that are democratic and with some type formers
(`finlim_biequiv_dfl_comp_cat_psfunctor` in the file `Biequiv`). There
is also some stuff about adjoint equivalences of comprehension
categories. Here univalence is used, because it contains multiple usages
of induction on equivalences.