reduce number of reversible coercions (#1859)
This partially undoes PR #1850.

Currently, it concerns 99 of the 259 reversibility declarations. I only
found 5 worthy to retain. I'll try to look at the others as well.

Unfortunately, this PR sits on the already merged PR #1858. So, only my
commits are relevant. The file comparison should be okay (not be
cluttered because of that).

Update: 250 of the 259 reversibility declarations are now gone. In one
single case was the need for reversibility circumvented by a small code
modification. The other 9 reversible coercions are now in use. If CI
goes well, this completes the present PR.