Extending the biequivalence by Clairambault and Dybjer to local properties (#1862)
This PR contains an extension of the biequivalence by Clairambault and
Dybjer to include local properties of categories. Currently, the example
of this is given by lextensive categories, so we get a biequivalence
between the bicategory of univalent lextensive categories and democratic
full comprehension categories with ∑, extensional identity, empty types,
and sum types.