Enriched profunctors (#1880)
This PR develops the basics of enriched profunctors. The main idea is to
use a whiskered presentation of profunctors, which is similar to how
monoidal categories and bicategories are developed. The reason for this
choice, is that the ultimate definition of profunctor becomes simpler
and more usable. For that purpose, whiskered and curried enriched
bifunctors are developed, and they are shown to be equivalent to the
usual definition.

Other stuff in this PR are transformations between enriched profunctors
(again in a whiskered style), and some of the main examples of enriched
profunctors and transformations between them.

---------

Co-authored-by: Benedikt Ahrens <benedikt.ahrens@gmail.com>