Remove outdated BUILD_COQ instructions (#1907)
(Background: The Makefile's `BUILD_COQ` option currently builds Coq
8.16.1, which is too old to compile UniMath; unfortunately, Coq's build
process completely changed as of Coq 8.17 so it is non-trivial to update
`BUILD_COQ` to a newer version of Coq.)

In the first draft of this PR, I have removed the mention of `BUILD_COQ`
from `INSTALL.md`. IMO, this is the minimal change that will keep users
from having to learn all of this the hard way (as I did). Additionally,
I notice that many of the install instructions haven't been updated in a
couple years and could be simplified by mentioning the new build system
(e.g., `INSTALL_OPAM.md` and `INSTALL_COQIDE.md`), and I am willing to
take a crack at some of this but I wanted to check in with the
developers first. (I am not interested in working on the Windows or Nix
instructions, though.)

A more ambitious project would be to actually remove the considerable
`BUILD_COQ` machinery from the build process altogether. I think this is
worth doing because it considerably simplifies the Makefile, removes the
useless git submodules, etc., but maybe that should be a different PR.